+++ src/Registry.php
@@ -2,11 +2,13 @@
 
 namespace Glhd\Gretel;
 
+use Closure;
 use Glhd\Gretel\Exceptions\MissingBreadcrumbException;
 use Glhd\Gretel\Routing\RouteBreadcrumb;
 use Illuminate\Routing\Route;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Traits\ForwardsCalls;
+use Throwable;
 
 /**
  * @mixin Collection
@@ -15,11 +17,74 @@ class Registry
 {
 	use ForwardsCalls;
 	
+	protected const HANDLER_MISSING = 'missing';
+	
+	protected const HANDLER_MISCONFIGURED = 'misconfigured';
+	
 	protected Collection $breadcrumbs;
+	
+	protected Collection $exception_handlers;
 	
 	public function __construct()
 	{
 		$this->breadcrumbs = new Collection();
+		$this->exception_handlers = new Collection();
+		
+		// Default to not throwing on missing breadcrumbs, only mis-configured ones
+		$this->throwOnMissingBreadcrumbs(false);
+	}
+	
+	public function withExceptionHandling(Closure $callback)
+	{
+		try {
+			return $callback();
+		} catch (MissingBreadcrumbException $exception) {
+			$this->callHandler(static::HANDLER_MISSING, $exception);
+		} catch (Throwable $exception) {
+			$this->callHandler(static::HANDLER_MISCONFIGURED, $exception);
+		}
+		
+		return null;
+	}
+	
+	public function handleMissingBreadcrumbs(Closure $callback): self
+	{
+		$this->exception_handlers->put(static::HANDLER_MISSING, $callback);
+		
+		return $this;
+	}
+	
+	public function throwOnMissingBreadcrumbs(bool $throw = true): self
+	{
+		if (!$throw) {
+			return $this->handleMissingBreadcrumbs(static function() {
+				// Ignore exception
+			});
+		}
+		
+		return $this->handleMissingBreadcrumbs(static function(Throwable $throwable) {
+			throw $throwable;
+		});
+	}
+	
+	public function handleMisconfiguredBreadcrumbs(Closure $callback): self
+	{
+		$this->exception_handlers->put(static::HANDLER_MISCONFIGURED, $callback);
+		
+		return $this;
+	}
+	
+	public function throwOnMisconfiguredBreadcrumbs(bool $throw = true): self
+	{
+		if (!$throw) {
+			return $this->handleMisconfiguredBreadcrumbs(static function() {
+				// Ignore exception
+			});
+		}
+		
+		return $this->handleMisconfiguredBreadcrumbs(static function(Throwable $throwable) {
+			throw $throwable;
+		});
 	}
 	
 	public function clear(): Registry
@@ -80,5 +145,16 @@ class Registry
 		}
 		
 		return $result;
+	}
+	
+	protected function callHandler(string $kind, Throwable $exception): void
+	{
+		$handler = $this->exception_handlers->get($kind);
+		
+		if (!$handler instanceof Closure) {
+			throw $exception;
+		}
+		
+		$handler($exception);
 	}
 }