+++ src/View/Components/Breadcrumbs.php
@@ -3,11 +3,11 @@
 namespace Glhd\Gretel\View\Components;
 
 use Glhd\Gretel\Exceptions\MissingBreadcrumbException;
-use Glhd\Gretel\Routing\RequestBreadcrumbs as RouteBreadcrumbs;
+use Glhd\Gretel\Registry;
+use Glhd\Gretel\Routing\RequestBreadcrumbs;
 use Glhd\Gretel\View\Breadcrumb;
 use Glhd\Gretel\View\BreadcrumbCollection;
 use Illuminate\Config\Repository;
-use Illuminate\Support\Facades\URL;
 use Illuminate\Support\HtmlString;
 use Illuminate\View\Component;
 
@@ -20,20 +20,24 @@ class Breadcrumbs extends Component
 	protected ?string $view = null;
 	
 	public function __construct(
-		RouteBreadcrumbs $breadcrumbs,
+		RequestBreadcrumbs $breadcrumbs,
+		Registry $registry,
 		Repository $config,
 		string $framework = null,
 		string $view = null,
 		bool $jsonLd = false,
-		bool $rdfa = false,
-		bool $throwIfMissing = false
+		bool $rdfa = false
 	) {
-		$this->breadcrumbs = $breadcrumbs->toCollection();
-		$this->config = $config;
+		// First, we'll initialize to an empty collection
+		$this->breadcrumbs = new BreadcrumbCollection();
 		
-		if ($throwIfMissing && $this->breadcrumbs->isEmpty()) {
-			throw new MissingBreadcrumbException(URL::current()); // FIXME
-		}
+		// Then we'll try to load the breadcrumbs, passing exceptions on to any configured handlers
+		$registry->withExceptionHandling(function() use ($breadcrumbs) {
+			$this->breadcrumbs = $breadcrumbs->toCollection();
+			$breadcrumbs->throwIfMissing();
+		});
+		
+		$this->config = $config;
 		
 		if ($view) {
 			$this->view = $view;
@@ -103,5 +107,14 @@ class Breadcrumbs extends Component
 			'@type' => 'BreadcrumbList',
 			'itemListElement' => $items,
 		], JSON_THROW_ON_ERROR | $flags));
+	}
+	
+	protected function getBreadcrumbCollection(RequestBreadcrumbs $breadcrumbs, Registry $registry): BreadcrumbCollection
+	{
+		try {
+			return $breadcrumbs->toCollection();
+		} catch (MissingBreadcrumbException $exception) {
+			// $registry->
+		}
 	}
 }